Add Linux port (CMake build, platform layer, runtime fixes)
Builds and runs Heretic2R natively on Linux. The engine, launcher, and all
game-side modules compile/link as ELF shared objects loaded via dlopen(), and
the game boots, loads levels, renders (ref_gl1/OpenGL), plays sound (SDL3) and
OGG music.

Build system & platform layer (src/linux/):
- CMakeLists.txt mirroring the 8-module MSVC solution layout.
- linux_compat.h: force-included Win32/MSVC shim (secure-CRT *_s functions with
  C++ array-overloads, min/max, HINSTANCE/DWORD, LoadLibrary->dlopen, __declspec,
  etc.) plus stub <windows.h>/<direct.h>/<io.h>.
- sys_unix.c / q_shunix.c / net_udp.c / launcher_main.c: system layer, main loop,
  filesystem/time helpers, BSD-sockets UDP, and the main() entry (replacing the
  Win32 WinMain / net_wins / q_shwin / sys_win).
- gen-case-symlinks.py: regenerates case-correct header symlinks (Linux is
  case-sensitive; not committed since git symlinks misbehave on Windows).

Source fixes (guarded for Windows where applicable):
- Hunk.c: VirtualAlloc/VirtualFree -> mmap/munmap.
- DllMain.c: Windows DllMain -> ELF constructor/destructor.
- Sys_dlopen uses RTLD_NODELETE: keeps modules mapped after dlclose so the
  load-probe-unload pattern (Client Effects, renderer/sound) doesn't leave a
  dangling API pointer (which on Linux hit reused address space and crashed).
- SinglyLinkedList.h: SLL_NODE_SIZE 8 (32-bit) -> 2*sizeof(void*); the node has
  a pointer-holding union, so it is 16 bytes on LP64 (was heap corruption).
- netmsg_write.c / Message.c: va_arg(byte|short) -> va_arg(int) + cast (the
  promotable-type read is UB; GCC compiled it to a trap).
- ce_Main.c: skip client-effects Clear() until Init() has set up its managers
  (early CLFX_Init on Linux ran Clear() before initialization).
- g_ClassStatics.h: define enum ClassID_e before its includes (GCC rejects an
  incomplete enum as a struct field; include cycle made it incomplete).
- g_Obj.h / ce_*.h / gl1_*.h: file-scope forward declarations (C tag scoping).
- qcommon.h: define BUILDSTRING/CPUSTRING for non-Windows.
- SurfaceProps.c: match the const void* prototype.
- gl1_FlexModel.c: replace a stray Latin-1 (0xF1) identifier byte with ASCII.
- p_items.c: drop in-initializer #pragma region markers (invalid in GCC).
- Scripting subsystem (game/src/Scripting/*.cpp) now compiled into gamex86.

See BUILDING-linux.md for build instructions and the full rationale.

+++ src/qcommon/Message.c
@@ -27,12 +27,12 @@ size_t MSG_SetParms(SinglyLinkedList_t* parms, const char* format, va_list marke
 		switch (format[count])
 		{
 			case 'b':
-				parm.t_byte = va_arg(marker, byte);
+				parm.t_byte = (byte)va_arg(marker, int); // Linux: va_arg promotes byte->int
 				bytesParsed += sizeof(parm.t_byte);
 				break;
 
 			case 's':
-				parm.t_short = va_arg(marker, short);
+				parm.t_short = (short)va_arg(marker, int); // Linux: va_arg promotes short->int
 				bytesParsed += sizeof(parm.t_short);
 				break;
